Homepage to list all pages and remove posts from header
This aligns with the original design. While the old hugo
is used the on home page all pages are listed. While upgrading to new
hugo, it looks the home page will not list all pages. This hack of
changes remove the 'posts' from header and list all pages on homepage.